refactor: inline imports plugin into vitest
